Phase 4.3: Move *_test.py into tests/ directory
Tests, test support code and test config fixtures now live in a top-level
tests/ package instead of alongside the implementation modules:

- moved *_test.py, test_lib.py, test_config_*.yaml and
  conflicting_test_config.yaml into tests/
- rewrote relative imports to absolute repo_structure.* imports
- dropped the ineffective `exclude = ["repo_structure/*_test.py"]` from
  packages.find (it filters package names, not files, so the old wheels
  did ship the tests) and pinned pytest discovery to testpaths = ["tests"]
- reworked repo_structure.yaml for the new layout

Closes #182
