+++ tests/test_config_with_warnings.yaml
